[breadboard-web] Massage harness into better shape. (#248)
- Add Breadbuddy to workspace.
- Move harness things to harness/.
- Teach ProxyReceiver to not be async gen.
- Get rid of async gen-related types.
- Fold Recever into Worker Harness.
- Clean up.
- Extrude createHarness factory.
- Teach MainThreadHarness about HarnessConfig.
- Add missing file.
- Teach ProxyServerHarness about HarnessConfig.
