+++ graphs.code-workspace
@@ -62,6 +62,9 @@
     },
     {
       "path": "packages/breadboard-cli"
+    },
+    {
+      "path": "packages/breadbuddy"
     }
   ],
   "settings": {}
