+++ packages/breadboard-web/src/harness/types.ts
@@ -4,6 +4,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import type {
+  Kit,
+  NodeTypeIdentifier,
+  OutputValues,
+} from "@google-labs/breadboard";
+
 export interface Harness {
   run(
     url: string,
@@ -21,3 +27,13 @@ export interface HarnessRunResult {
   reply(reply: unknown): void;
   message: Result;
 }
+
+export type SecretHandler = (keys: {
+  keys?: string[];
+}) => Promise<OutputValues>;
+
+export type HarnessConfig = {
+  proxy: NodeTypeIdentifier[];
+  kits: Kit[];
+  onSecret: SecretHandler;
+};
